+++ inst/include/dplyr.h
@@ -57,6 +57,7 @@ namespace dplyr {
     template <typename Index>
     DataFrame subset( DataFrame df, const Index& indices, CharacterVector classes) ;
     void check_attribute_compatibility( SEXP left, SEXP right) ;
+    bool same_levels( SEXP left, SEXP right ) ;
 }
 dplyr::Result* get_handler( SEXP, const dplyr::LazySubsets&, const Environment& ) ;
 dplyr::Result* nth_prototype( SEXP call, const dplyr::LazySubsets& subsets, int nargs) ;