+++ tests/testthat/test-sets.R
@@ -54,5 +54,20 @@ test_that("set operations use coercion rules (#799)", {
   expect_equal( res, data_frame(x = letters[1:5]) )
   expect_warning( { res <- setdiff(df2, df1) })
   expect_equal( res, data_frame(x = letters[11:15]) )
+})
 
+test_that("setdiff handles factors with NA (#1526)", {
+  df1 <- data_frame(x = factor(c(NA, "a")))
+  df2 <- data_frame(x = factor("a"))
+
+  res <- setdiff(df1, df2)
+  expect_is( res$x, "factor")
+  expect_equal( levels(res$x), "a")
+  expect_true( is.na(res$x[1]) )
+
+
+  df1 <- data_frame(x = factor(c(NA, "b")))
+  expect_warning( res <- setdiff(df1, df2) )
+  expect_is( res$x, "character")
+  expect_true( is.na(res$x[1]) )
 })